✨ feat: map WPF activity stages to DateTime expire windows
Split `migrate/wpf.rs` into per-task modules under `migrate/wpf/`.
Read side-story expire time from StageActivityV2.json at migrate time
instead of emitting OnSideStory; fail after hot-update if the activity
file cannot be read. Rename migrate `--config` to `--profile-name`.

将 WPF 活动关迁移为 DateTime 截止时间，并拆分 migrate/wpf 模块。
