+++ crates/maa-cli/src/main.rs
@@ -111,12 +111,12 @@ fn main() -> Result<()> {
         Command::Migrate(command::MigrateCommand::Wpf {
             input,
             output,
-            config,
+            profile_name,
             custom_schedule,
         }) => config::migrate::wpf(
             &input,
             output.as_deref(),
-            config,
+            profile_name,
             custom_schedule.as_deref(),
         )?,
         Command::Activity { client } => activity::display_stage_activity(client)?,
